// apps/frontend/src/lib/api/generated/models/SeoMetaPublic.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
/**
 * Public serializer for SEO metadata (for frontend)
 */
export type SeoMetaPublic = {
    /**
     * home, city, service, city_service, about, contacts
     */
    page_type: string;
    title: string;
    meta_description: string;
    meta_keywords?: string;
    h1_title: string;
    full_slug: string;
    og_title?: string;
    og_description?: string;
    og_image_url?: string;
    schema_json?: any;
};